requirements.txt
langchain
langchain-openai
langgraph
langfuse
python-dotenv
supabase
openai

# For evaluation
nltk
rouge
bert-score
scikit-learn # For BertScore dependencies
konlpy